Fix one mul-overflow issue
When minInterval is set to 5s or larger,
mul-overflow will happen when we convert
minInterval from second to nano second in
getElapsedRealtimeEstimateNanos function.
Change the type of time between fix parameter
to int64_t.

CRs-Fixed: 2848319

+++ core/LocApiBase.cpp
@@ -930,7 +930,7 @@ void LocApiBase::
 DEFAULT_IMPL()
 
 int64_t ElapsedRealtimeEstimator::getElapsedRealtimeEstimateNanos(int64_t curDataTimeNanos,
-            bool isCurDataTimeTrustable, uint32_t tbf) {
+            bool isCurDataTimeTrustable, int64_t tbf) {
     //The algorithm works follow below steps:
     //When isCurDataTimeTrustable is meet (means Modem timestamp is already stable),
     //1, Wait for mFixTimeStablizationThreshold fixes; While waiting for modem time
